requirements.txt
fastapi[standart]
uvicorn
python-dotenv